fix(chat-input): support Ctrl+Z undo for pasted images and hide placeholder
